fix: clippy warnings

+++ packages/wasm-dpp/src/document/state_transition/batch_transition/mod.rs
@@ -105,7 +105,7 @@ impl BatchTransitionWasm {
 
     #[wasm_bindgen(js_name=getUserFeeIncrease)]
     pub fn get_user_fee_increase(&self) -> u16 {
-        self.0.user_fee_increase() as u16
+        self.0.user_fee_increase()
     }
 
     #[wasm_bindgen(js_name=setUserFeeIncrease)]
